--- tests/baselines/reference/destructuringParameterDeclaration10.symbols
@@ -1,30 +1,0 @@
-//// [tests/cases/conformance/es6/destructuring/destructuringParameterDeclaration10.ts] ////
-
-=== destructuringParameterDeclaration10.ts ===
-const getUserProfileByUsername = ({
->getUserProfileByUsername : Symbol(getUserProfileByUsername, Decl(destructuringParameterDeclaration10.ts, 0, 5))
-
-  username,
->username : Symbol(username, Decl(destructuringParameterDeclaration10.ts, 0, 35))
-
-  withEmail = false,
->withEmail : Symbol(withEmail, Decl(destructuringParameterDeclaration10.ts, 1, 11))
-
-  withNotificationSettings = false,
->withNotificationSettings : Symbol(withNotificationSettings, Decl(destructuringParameterDeclaration10.ts, 2, 20))
-
-} = {}) => {
-  return {};
-};
-
-getUserProfileByUsername({
->getUserProfileByUsername : Symbol(getUserProfileByUsername, Decl(destructuringParameterDeclaration10.ts, 0, 5))
-
-  username: 'ly', // ok (implicit any reported at declaration site)
->username : Symbol(username, Decl(destructuringParameterDeclaration10.ts, 8, 26))
-
-});
-
-getUserProfileByUsername({}); // ok
->getUserProfileByUsername : Symbol(getUserProfileByUsername, Decl(destructuringParameterDeclaration10.ts, 0, 5))
-
