--- tests/baselines/reference/destructuringParameterDeclaration9.symbols
@@ -1,34 +1,0 @@
-//// [tests/cases/conformance/es6/destructuring/destructuringParameterDeclaration9.ts] ////
-
-=== utils.js ===
-export const getUserProfileByUsername = ({
->getUserProfileByUsername : Symbol(getUserProfileByUsername, Decl(utils.js, 0, 12))
-
-  username,
->username : Symbol(username, Decl(utils.js, 0, 42))
-
-  withEmail = false,
->withEmail : Symbol(withEmail, Decl(utils.js, 1, 11))
-
-  withNotificationSettings = false,
->withNotificationSettings : Symbol(withNotificationSettings, Decl(utils.js, 2, 20))
-
-} = {}) => {
-  return {};
-};
-
-=== main.ts ===
-import { getUserProfileByUsername } from './utils';
->getUserProfileByUsername : Symbol(getUserProfileByUsername, Decl(main.ts, 0, 8))
-
-getUserProfileByUsername({
->getUserProfileByUsername : Symbol(getUserProfileByUsername, Decl(main.ts, 0, 8))
-
-  username: 'ly', // ok
->username : Symbol(username, Decl(main.ts, 2, 26))
-
-});
-
-getUserProfileByUsername({}); // ok
->getUserProfileByUsername : Symbol(getUserProfileByUsername, Decl(main.ts, 0, 8))
-
