--- tests/baselines/reference/destructuringParameterDeclaration9.types
@@ -1,64 +1,0 @@
-//// [tests/cases/conformance/es6/destructuring/destructuringParameterDeclaration9.ts] ////
-
-=== utils.js ===
-export const getUserProfileByUsername = ({
->getUserProfileByUsername : ({ username, withEmail, withNotificationSettings, }?: { username?: any; withEmail?: boolean; withNotificationSettings?: boolean; }) => {}
->                         : ^                                                  ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
->({  username,  withEmail = false,  withNotificationSettings = false,} = {}) => {  return {};} : ({ username, withEmail, withNotificationSettings, }?: { username?: any; withEmail?: boolean; withNotificationSettings?: boolean; }) => {}
->                                                                                              : ^                                                  ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
-
-  username,
->username : any
->         : ^^^
-
-  withEmail = false,
->withEmail : boolean
->          : ^^^^^^^
->false : false
->      : ^^^^^
-
-  withNotificationSettings = false,
->withNotificationSettings : boolean
->                         : ^^^^^^^
->false : false
->      : ^^^^^
-
-} = {}) => {
->{} : {}
->   : ^^
-
-  return {};
->{} : {}
->   : ^^
-
-};
-
-=== main.ts ===
-import { getUserProfileByUsername } from './utils';
->getUserProfileByUsername : ({ username, withEmail, withNotificationSettings, }?: { username?: any; withEmail?: boolean; withNotificationSettings?: boolean; }) => {}
->                         : ^                                                  ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
-
-getUserProfileByUsername({
->getUserProfileByUsername({  username: 'ly', // ok}) : {}
->                                                    : ^^
->getUserProfileByUsername : ({ username, withEmail, withNotificationSettings, }?: { username?: any; withEmail?: boolean; withNotificationSettings?: boolean; }) => {}
->                         : ^                                                  ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
->{  username: 'ly', // ok} : { username: string; }
->                          : ^^^^^^^^^^^^^^^^^^^^^
-
-  username: 'ly', // ok
->username : string
->         : ^^^^^^
->'ly' : "ly"
->     : ^^^^
-
-});
-
-getUserProfileByUsername({}); // ok
->getUserProfileByUsername({}) : {}
->                             : ^^
->getUserProfileByUsername : ({ username, withEmail, withNotificationSettings, }?: { username?: any; withEmail?: boolean; withNotificationSettings?: boolean; }) => {}
->                         : ^                                                  ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
->{} : {}
->   : ^^
-
