+++ tests/cases/conformance/es6/destructuring/destructuringParameterDeclaration12.ts
